perf: 64-bit decompression bitstream with 4-byte absorption and bulk RLE fill
Widen DState.bsBuff from UInt32 to UInt64 and modify the GET_BITS macro
to absorb 4 bytes at once when input is available and the buffer has
room (bsLive <= 32). This reduces the absorption loop iterations from
3-4 down to 1 in the hot Huffman decode path. At BZ_STREAM_END, unused
whole bytes are returned to the input stream to preserve correct
concatenated stream handling.

Also hoist the bounds check out of the RUNA/RUNB fill loop so the
compiler can optimize the inner fill without per-element branching.

+++ src/qbz2_internal.h
@@ -313,7 +313,7 @@ typedef
       Bool     blockRandomised;
       BZ_RAND_DECLS;
 
-      UInt32   bsBuff;
+      UInt64   bsBuff;
       Int32    bsLive;
 
       Int32    blockSize100k;
